Splitted VideoDetailView
Added new components (Tags, Stats, Description), to simplify readability/logic inside the VideoDetailView component
